MadGraphReweightingProducer.cc: minor change to be safe

+++ src/Producers/MadGraphReweightingProducer.cc
@@ -116,7 +116,7 @@ void MadGraphReweightingProducer::Init(setting_type const& settings)
 	});
 	LambdaNtupleConsumer<HttTypes>::AddIntQuantity("lheParticleJetNumber", [](event_type const& event, product_type const& product)
 	{
-		return product.m_lheParticlesSortedForMadGraph.size() - 3;
+		return static_cast<int>(product.m_lheParticlesSortedForMadGraph.size()) - 3;
 	});
 }
